board, list 기초 파일 생성
