--- src/main/java/com/example/trello/board/controller/BoardController.java
@@ -1,4 +1,0 @@
-package com.example.trello.board.controller;
-
-public class BoardController {
-}
