--- src/main/java/com/example/trello/board/dto/BoardRequestDto.java
@@ -1,4 +1,0 @@
-package com.example.trello.board.dto;
-
-public class BoardRequestDto {
-}
