--- src/main/java/com/example/trello/board/dto/BoardResponseDto.java
@@ -1,4 +1,0 @@
-package com.example.trello.board.dto;
-
-public class BoardResponseDto {
-}
