--- src/main/java/com/example/trello/board/entity/Board.java
@@ -1,4 +1,0 @@
-package com.example.trello.board.entity;
-
-public class Board {
-}
