--- src/main/java/com/example/trello/board/repository/BoardRepository.java
@@ -1,4 +1,0 @@
-package com.example.trello.board.repository;
-
-public interface BoardRepository {
-}
