--- src/main/java/com/example/trello/board/service/BoardService.java
@@ -1,4 +1,0 @@
-package com.example.trello.board.service;
-
-public class BoardService {
-}
