+++ src/main/java/com/example/trello/domain/board/controller/BoardController.java
@@ -1,0 +1,4 @@
+package com.example.trello.domain.board.controller;
+
+public class BoardController {
+}
