+++ src/main/java/com/example/trello/domain/board/dto/BoardResponseDto.java
@@ -1,0 +1,4 @@
+package com.example.trello.domain.board.dto;
+
+public class BoardResponseDto {
+}
