+++ src/main/java/com/example/trello/domain/board/repository/BoardRepository.java
@@ -1,0 +1,4 @@
+package com.example.trello.domain.board.repository;
+
+public interface BoardRepository {
+}
