+++ src/main/java/com/example/trello/domain/board/service/BoardService.java
@@ -1,0 +1,4 @@
+package com.example.trello.domain.board.service;
+
+public class BoardService {
+}
