+++ src/main/java/com/example/trello/domain/list/controller/ListController.java
@@ -1,0 +1,4 @@
+package com.example.trello.domain.list.controller;
+
+public class ListController {
+}
