+++ src/main/java/com/example/trello/domain/list/dto/ListRequestDto.java
@@ -1,0 +1,4 @@
+package com.example.trello.domain.list.dto;
+
+public class ListRequestDto {
+}
