+++ src/main/java/com/example/trello/domain/list/repository/ListRepository.java
@@ -1,0 +1,4 @@
+package com.example.trello.domain.list.repository;
+
+public interface ListRepository {
+}
