+++ src/main/java/com/example/trello/domain/list/service/ListService.java
@@ -1,0 +1,4 @@
+package com.example.trello.domain.list.service;
+
+public class ListService {
+}
