--- src/main/java/com/example/trello/list/controller/ListController.java
@@ -1,4 +1,0 @@
-package com.example.trello.list.controller;
-
-public class ListController {
-}
