--- src/main/java/com/example/trello/list/dto/ListRequestDto.java
@@ -1,4 +1,0 @@
-package com.example.trello.list.dto;
-
-public class ListRequestDto {
-}
