--- src/main/java/com/example/trello/list/dto/ListResponseDto.java
@@ -1,4 +1,0 @@
-package com.example.trello.list.dto;
-
-public class ListResponseDto {
-}
