--- src/main/java/com/example/trello/list/entity/List.java
@@ -1,4 +1,0 @@
-package com.example.trello.list.entity;
-
-public class List {
-}
