--- src/main/java/com/example/trello/list/repository/ListRepository.java
@@ -1,4 +1,0 @@
-package com.example.trello.list.repository;
-
-public interface ListRepository {
-}
