--- src/main/java/com/example/trello/list/service/ListService.java
@@ -1,4 +1,0 @@
-package com.example.trello.list.service;
-
-public class ListService {
-}
